requirements.txt
Pillow>=10.0.0
markdown>=3.5.0
jinja2>=3.1.0
pdfkit>=0.6.1
